requirements.txt
# AGNTCY Multi-Organization Image Classification System
# Python 3.12+ required (for a2a-sdk and agntcy-app-sdk)

# Web frameworks
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiohttp>=3.11.0

# Data validation (a2a-sdk requires pydantic>=2.11.3)
pydantic>=2.11.3
pydantic-settings>=2.7.0

# Transport & messaging (A2A SDK - exact versions required)
nats-py>=2.6.0
a2a-sdk==0.3.2
agntcy-app-sdk==0.4.6

# Object storage
minio>=7.2.3

# Image processing
Pillow>=10.2.0
numpy>=1.26.3

# Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-exporter-otlp>=1.22.0

# Utilities
python-multipart>=0.0.6

# LangGraph & LangChain (for AI agent workflows)
langgraph>=0.2.45
langchain>=0.3.7
langchain-core>=0.3.15
langchain-openai>=0.2.8
langchain-anthropic>=0.2.4

# LLM Integration (Lungo-style)
litellm>=1.55.0
llama-index>=0.12.0
llama-index-llms-litellm>=0.3.0
llama-index-llms-openai>=0.3.0
llama-index-llms-anthropic>=0.3.0
